requirements_refactored.txt
websockets>=12.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
google-generativeai>=0.8.0
wave
